plugins/umap_plot/requirements.txt
# Dependencies für UMAP Plot Plugin
umap-learn>=0.5.0
plotly>=5.0.0
pandas>=1.3.0
numpy>=1.20.0